Added fullscreen and windowed mode toggling for Linux
The Lua functions fullscreenMode() and windowedMode(x,y,w,h)
now work as you would expect for Linux. The Windows version
was already working.

The function glutFullscreen was corrected.
Code was added to glutPositionWindow so it switches back to windowed mode.

Removed the initial glutFullscreen call. Praxis starts windowed by default.

Corrected luaCBWinGetScreenWidth and luaCBWinGetScreenHeight so
moveWindowFull has the borders going neatly to the edge of the screen.